Week2/Code_demo/Study/requirements.txt
flask==3.0.0
requests==2.31.0
PyJWT==2.8.0
requests
